Move TestExporter to avoid requiring pytest
